+++ src/main/java/ch/metzenthin/svm/domain/commands/DeleteKurstypCommand.java
@@ -31,10 +31,10 @@ public class DeleteKurstypCommand extends GenericDaoCommand {
     public void execute() {
         KurstypDao kurstypDao = new KurstypDao(entityManager);
         Kurstyp kurstypToBeDeleted = kurstypen.get(indexKurstypToBeDeleted);
-//        if (kurstypToBeDeleted.getKurse().size() > 0) {
-//            result = Result.KURSTYP_VON_KURS_REFERENZIERT;
-//            return;
-//        }
+        if (kurstypToBeDeleted.getKurse().size() > 0) {
+            result = Result.KURSTYP_VON_KURS_REFERENZIERT;
+            return;
+        }
         kurstypDao.remove(kurstypToBeDeleted);
         kurstypen.remove(indexKurstypToBeDeleted);
         result = Result.LOESCHEN_ERFOLGREICH;
